+++ collections/m2m gateway/eservices/Create eservice risk analyses.bru
@@ -4,7 +4,7 @@ meta {
 }
 
 post {
-  url: {{host-m2m-gw}}/eservices/:eserviceId/riskAnalysis
+  url: {{host-m2m-gw}}/eservices/:eserviceId/riskAnalyses
   body: json
   auth: none
 }
