+++ collections/m2m gateway/eservices/Delete eservice risk analysis.bru
@@ -4,7 +4,7 @@ meta {
 }
 
 delete {
-  url: {{host-m2m-gw}}/eservices/:eserviceId/riskAnalysis/:riskAnalysisId
+  url: {{host-m2m-gw}}/eservices/:eserviceId/riskAnalyses/:riskAnalysisId
   body: none
   auth: none
 }
