+++ collections/m2m gateway/eservices/Get eservice risk analysis.bru
@@ -1,10 +1,10 @@
 meta {
-  name: Get eservice risk analysis
+  name: Get eservice risk analyses
   type: http
 }
 
 get {
-  url: {{host-m2m-gw}}/eservices/:eserviceId/riskAnalysis/:riskAnalysisId
+  url: {{host-m2m-gw}}/eservices/:eserviceId/riskAnalyses/:riskAnalysisId
   body: none
   auth: none
 }
