+++ collections/m2m gateway/eservices/List eservice risk analyses.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{host-m2m-gw}}/eservices/:eserviceId/riskAnalysis?offset=0&limit=10
+  url: {{host-m2m-gw}}/eservices/:eserviceId/riskAnalyses?offset=0&limit=10
   body: none
   auth: none
 }
