+++ packages/api-clients/open-api/m2mGatewayApi.yml
@@ -8383,7 +8383,7 @@ paths:
       summary: Create a new Risk Analysis
       description: |
         Create a Risk Analysis for an E-Service Template.
-        The E-Service Template must be in Draft state.
+        The E-Service Template must be in Draft state and in Receive mode.
       operationId: createEServiceTemplateRiskAnalysis
       requestBody:
         required: true
