+++ packages/m2m-gateway/test/api/eServiceTemplates/CreateEserviceTemplateRiskAnalysis.test.ts
@@ -49,7 +49,7 @@ describe("POST /eserviceTemplates/:templateId/riskAnalyses router test", () => {
 
   const authorizedRoles: AuthRole[] = [authRole.M2M_ADMIN_ROLE];
   it.each(authorizedRoles)(
-    "Should return 201 and perform service calls for user with role %s",
+    "Should return 200 and perform service calls for user with role %s",
     async (role) => {
       mockEServiceTemplateService.createEServiceTemplateRiskAnalysis = vi
         .fn()
@@ -62,7 +62,7 @@ describe("POST /eserviceTemplates/:templateId/riskAnalyses router test", () => {
         mockRiskAnalysisSeed
       );
 
-      expect(res.status).toBe(201);
+      expect(res.status).toBe(200);
       expect(res.body).toEqual(mockM2MEserviceResponse);
     }
   );
